Do you really think I'm going to write out everything I changed? Here's a brief overview: Animations App List Download Progress Uninstall Modal alerts Colors Icons Preferences Search Bar

+++ Redstone2/Core/RSMetrics.m
@@ -4,7 +4,7 @@
 
 int columns = 3;
 
-+ (NSInteger)columns {
++ (int)columns {
 	return columns;
 }
